Simplify state management by using useEffect
