+++ airflow/ui/src/components/ClearRun/ClearRunDialog.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { Flex, Heading, VStack } from "@chakra-ui/react";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { FiRefreshCw } from "react-icons/fi";
 
 import type {
@@ -45,44 +45,33 @@ type Props = {
     requestBody: DAGRunClearBody;
   }) => void;
   readonly onClose: () => void;
-  readonly onlyFailed: boolean;
   readonly open: boolean;
-  readonly setOnlyFailed: (value: boolean) => void;
 };
 
-const ClearRunDialog = ({
-  affectedTasks,
-  dagRun,
-  isPending,
-  mutate,
-  onClose,
-  onlyFailed,
-  open,
-  setOnlyFailed,
-}: Props) => {
+const ClearRunDialog = ({ affectedTasks, dagRun, isPending, mutate, onClose, open }: Props) => {
+  const [onlyFailed, setOnlyFailed] = useState(false);
+
   const dagId = dagRun.dag_id;
   const dagRunId = dagRun.dag_run_id;
 
   const [note, setNote] = useState<string | null>(dagRun.note);
   const { isPending: isPendingPatchDagRun, mutate: mutatePatchDagRun } = usePatchDagRun({ dagId, dagRunId });
 
+  useEffect(() => {
+    mutate({
+      dagId,
+      dagRunId,
+      requestBody: { dry_run: true, only_failed: onlyFailed },
+    });
+  }, [dagId, dagRunId, mutate, onlyFailed]);
+
   const onChange = (value: string) => {
     switch (value) {
       case "existing_tasks":
         setOnlyFailed(false);
-        mutate({
-          dagId,
-          dagRunId,
-          requestBody: { dry_run: true, only_failed: false },
-        });
         break;
       case "only_failed":
         setOnlyFailed(true);
-        mutate({
-          dagId,
-          dagRunId,
-          requestBody: { dry_run: true, only_failed: true },
-        });
         break;
       default:
         // TODO: Handle this `new_tasks` case
